### 2022-04-15 #525 @bcjaeger
Pre-submission Inquiry: aorsf; accelerated oblique random survival forests
Submitting Author Name: Byron C. Jaeger
Submitting Author Github Handle: @bcjaeger
Other Package Authors Github handles: (comma separated, delete if none) @nmpieyeskey, @sawyerWeld
Repository:  https://github.com/bcjaeger/aorsf
Submission type: Pre-submission

---

```
Package: aorsf
Title: Accelerated Oblique Random Survival Forests
Version: 0.0.0.9000
    person(given = "Byron",
           family = "Jaeger",
           email = "bjaeger@wakehealth.edu",
    person(given = "Nicholas",  family = "Pajewski", role = "ctb"),
    person(given = "Sawyer", family = "Welden", role = "ctb", email = "swelden@wakehealth.edu")
    )
Description: Fit, interpret, and make predictions with oblique random
    survival forests. Oblique decision trees are notoriously slow compared
    to their axis based counterparts, but 'aorsf' runs as fast or faster than
    axis-based decision tree algorithms for right-censored time-to-event
    outcomes.
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    Rcpp,
    RcppArmadillo
    table.glue,
    Rcpp,
    data.table
    https://bcjaeger.github.io/aorsf
    survival,
    survivalROC,
    glmnet,
    covr,
    units

```

	- [ ] text data

Random forests are a machine learning algorithm and this package provides optimized code to fit a specific type of random forest. I am unsure about whether this belongs in the machine learning category or the regression and supervised learning category. I am uncertain about whether `aorsf` belongs in regression and supervised learning because random forests are definitely used for supervised learning but they don't really fit into a 'regression' framework.

Yes

Target audience: people who want to develop or interpret a risk prediction model, i.e., a prediction model for right-censored time-to-event outcomes.

Applications: fit an oblique random survival forest, compute predicted risk at a given time, estimate the importance of individual variables, and compute partial dependence to depict relationships between specific predictors and predicted risk.

The `obliqueRSF` package precedes `aorsf`. The `aorsf` package runs hundreds of times faster than `obliqueRSF` and includes novel features for interpretation of the oblique random survival forest (negation importance and ANOVA importance). I developed both packages.

Yes.

None.
