+++ vignettes/pd.Rmd
@@ -12,7 +12,9 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>", 
+  fig.height = 5, 
+  fig.width = 7
 )
 ```
 
@@ -35,20 +37,20 @@ First fit an accelerated ORSF model on `pbc_orsf`, a modification of the Mayo Cl
 
 set.seed(123)
 
-osrf_fit <- orsf(data_train = pbc_orsf,
+orsf_fit <- orsf(data_train = pbc_orsf,
                  formula = Surv(time, status) ~ . -id,
                  n_tree = 2500)
 
 ```
 
-To view the partial dependence between blood vessel malformations in the skin (`spiders`) and urine copper (`copper`), use `orsf_pd_summary`. The `pd_spec` argument contains the variable values we want to compute partial dependence for. By default `oobag == TRUE`, meaning partial dependence is calculated using the out of bag training data. We keep this set to `TRUE` because we are computing partial dependence using the training data. `expand_grid == TRUE` tells `orsf_pd_summary` to use all combinations of values in `pd_spec`.
+To view the partial dependence between predicted risk and the combination of blood vessel malformations in the skin (`spiders`) and urine copper (`copper`), we use `orsf_pd_summary`. The `pd_spec` argument contains the variable values we want to compute partial dependence for. By default `oobag == TRUE`, meaning partial dependence is calculated using the out of bag training data. We keep this set to `TRUE` because we are computing partial dependence using the training data. `expand_grid == TRUE` tells `orsf_pd_summary` to use all combinations of values in `pd_spec`.
 
 ```{r}
 pd_spec <- list(spiders = c("0", "1"),
                 copper = seq(45, 130, by = 10))
 
-pd_data <- orsf_pd_summary(object = osrf_fit,
-                           pd_spec = pd_spec,,
+pd_data <- orsf_pd_summary(object = orsf_fit,
+                           pd_spec = pd_spec,
                            oobag = TRUE,
                            expand_grid = TRUE)
 
@@ -57,23 +59,25 @@ ggplot(pd_data) +
  geom_line() + 
  labs(y = 'Predicted risk',
       x = 'Urine Copper',
-      title = 'Partial dependence of copper and treatment')
+      title = 'Partial dependence of copper and spiders')
 ```
 
-Next lets look at Individual Conditional Expectation (ICE). Each instance in the dataset is represented by one line showing change in prediction dependant on change in a feature. While `orsf_pd_ice` allows you to use multiple features, it is recommended to view one at a time. In the plot below we can see how each datapoint's predicted survival changes as copper changes. 
+Next lets look at Individual Conditional Expectation (ICE). Each instance in the dataset is represented by one line showing change in prediction dependent on change in a feature. While `orsf_pd_ice` allows you to use multiple features, it is recommended to view one at a time. In the plot below we can see how each data point's predicted survival changes as copper changes. 
 
 
 ```{r}
-ice_data <- orsf_pd_ice(object = osrf_fit,
-                       pd_spec = list(bili = seq(1, 4, by = 0.5)),
+
+ice_data <- orsf_pd_ice(object = orsf_fit,
+                       pd_spec = list(copper = seq(45, 130, by = 10)),
                        expand_grid = TRUE)
 
-ice_data |>
- ggplot(aes(x=bili, y=pred, group=id_row)) + 
+ice_data[id_row %in% c(1:30)] |> 
+ ggplot(aes(x=copper, y=pred, group=id_row)) + 
  geom_line() +
  labs(y = 'Predicted risk',
      x = 'Urine Copper',
[…]
