### 2022-04-24 7c02fca
more lenient on time check to avoid spurious CI fails
