+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   aorsf: aorsf.html
   oobag: oobag.html
   pd: pd.html
-last_built: 2022-04-24T12:55Z
+last_built: 2022-04-24T14:25Z
 urls:
   reference: https://bcjaeger.github.io/aorsf/reference
   article: https://bcjaeger.github.io/aorsf/articles
