### 2022-04-26 #525 @jooolia
Hi @bcjaeger,
Thanks for your patience. It seems that the machine learning category would be a good fit for your package and the package  appears in good shape to make a full submission when you would like.

Regarding the categories, @mpadge may have a bit more to add about this.

Thanks, Julia
