### 2022-04-27 #525 @mpadge
Thanks for your submission @bcjaeger Our statistical standards are a work-in-progress. Please help to improve them by providing feedback, particularly on appropriateness or otherwise of any particular standard. That can be done informally via [GitHub discussions](https://github.com/ropensci/statistical-software-review-book/discussions), or more formally via a pull request to that same repo ([standards are here](https://github.com/ropensci/statistical-software-review-book/tree/main/standards)). We're also very keen to develop policies on handling cases of potentially ambiguous categories, such as yours. To help that process, I've started [this discussion thread](https://github.com/ropensci/statistical-software-review-book/discussions/63) - please offer any insight you can. Thanks!
